[FOLD] Move KeyType from crypto/ to protocol/
